+++ server/api_rpc.go
@@ -15,16 +15,13 @@
 package server
 
 import (
-	"net"
 	"strings"
 
 	"github.com/gofrs/uuid"
 	"github.com/heroiclabs/nakama/api"
-	"go.uber.org/zap"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
-	"google.golang.org/grpc/peer"
 	"google.golang.org/grpc/status"
 )
 
@@ -64,26 +61,8 @@ func (s *ApiServer) RpcFunc(ctx context.Context, in *api.Rpc) (*api.Rpc, error) 
 	if e := ctx.Value(ctxExpiryKey{}); e != nil {
 		expiry = e.(int64)
 	}
-	clientAddr := ""
-	clientIP := ""
-	clientPort := ""
-	md, _ := metadata.FromIncomingContext(ctx)
-	if ips := md.Get("x-forwarded-for"); len(ips) > 0 {
-		// look for gRPC-Gateway / LB header
-		clientAddr = strings.Split(ips[0], ",")[0]
-	} else if peerInfo, ok := peer.FromContext(ctx); ok {
-		// if missing, try to look up gRPC peer info
-		clientAddr = peerInfo.Addr.String()
-	}
-	clientAddr = strings.TrimSpace(clientAddr)
-	if host, port, err := net.SplitHostPort(clientAddr); err == nil {
-		clientIP = host
-		clientPort = port
-	} else if addrErr, ok := err.(*net.AddrError); ok && addrErr.Err == "missing port in address" {
-		clientIP = clientAddr
-	} else {
-		s.logger.Debug("Could not extract client address from request.", zap.Error(err))
-	}
+
+	clientIP, clientPort := extractClientAddressFromContext(s.logger, ctx)
 
 	result, fnErr, code := fn(ctx, queryParams, uid, username, expiry, "", clientIP, clientPort, in.Payload)
 	if fnErr != nil {